Backend/requirements.txt
django
djangorestframework
flask
